ADDED: image_box element
Import a picture into asciio through the system clipboard or picture file open
to become an image box. The `xclip` binary tool needs to be installed on Linux
systems.

ADDED: Element freeze attribute

If an element has the frozen attribute, the element cannot be resized or
selected. This element exists as a background in the current canvas.

If you want to operate on the element again, you need to set the
`IGNORE_FROZEN_ELEMENTS` attribute. Frozen elements are temporarily unlocked,
allowing resizing and selection.

The borders of frozen elements appear gray or blue depending on the theme.
If the `IGNORE_FROZEN_ELEMENTS` attribute is present, frozen elements
display a green border.

ADDED: App::Asciio::GTK::Asciio::Actions::File module.

Added open_image sub, used to open an image from a file into an
image box and insert it into the canvas.