WIP: image_box and element freeze attribute
ADDED: image_box element

Import a picture into asciio through the system clipboard to become an
image box. The `xclip` binary tool needs to be installed on Linux systems.

ADDED: Element freeze attribute

If an element has the frozen attribute, the element cannot be resized or
selected. This element exists as a background in the current canvas.

If you want to operate on the element again, you need to set the
`IGNORE_ELEMENT_FREEZE` attribute. Frozen elements are temporarily unlocked,
allowing resizing and selection.

The borders of frozen elements appear gray or blue depending on the theme.
If the `IGNORE_ELEMENT_FREEZE` attribute is present, frozen elements
display a green border.